requirements.txt
pvporcupine~=1.9.0
PyAudio~=0.2.11
paho-mqtt~=1.5.0
wit~=6.0.0
pyttsx3==2.7
pytz~=2019.3
geopy~=2.1.0
timezonefinder~=5.2.0
requests~=2.24.0
wikipedia~=1.4.0
importlib~=1.0.4
playsound~=1.2.2
spidev